% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries.R
\name{ConvertBboxToSpatialPolygon}
\alias{ConvertBboxToSpatialPolygon}
\title{Bbox to SpatialPolygons object}
\usage{
ConvertBboxToSpatialPolygon(bbox, proj4string)
}
\arguments{
\item{bbox}{Bounding box object}

\item{proj4string}{projection string of class CRS-class}
}
\description{
Bbox to SpatialPolygons object
}

