% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries.R
\name{CreateCorrespondingSpatialPoints}
\alias{CreateCorrespondingSpatialPoints}
\title{Returns appropriate SpatialPoints obeject for lon lat coords}
\usage{
CreateCorrespondingSpatialPoints(lon, lat, proj4string)
}
\arguments{
\item{lon}{numeric vector of longitudes}

\item{lat}{numeric vector of latitudes}

\item{proj4string}{projection string of class CRS-class (used by the area)}
}
\description{
Returns appropriate SpatialPoints obeject for lon lat coords
}

