% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries.R
\name{FindClosestGridPoints}
\alias{FindClosestGridPoints}
\title{Return closest grid point}
\usage{
FindClosestGridPoints(pointData, gridData, k = 1L)
}
\arguments{
\item{pointData}{Data.table containing lon and lat of the point of interest}

\item{gridData}{Data.table containing lon, lat, and pointID of the grid}

\item{k}{Integer How many points should be returned}
}
\description{
Return closest grid point
}

