% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries.R
\name{gClip}
\alias{gClip}
\title{Clip spatial geometry to bounding box}
\usage{
gClip(shp, bb)
}
\arguments{
\item{shp}{Spatial geometry}

\item{bb}{bounding box}
}
\description{
obtained from Robin Lovelace \url{http://robinlovelace.net/r/2014/07/29/clipping-with-r.html}
}

