% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{GetBoxCoxAdjustment}
\alias{GetBoxCoxAdjustment}
\title{Prepare for Box-Cox transformation}
\usage{
GetBoxCoxAdjustment(data)
}
\arguments{
\item{data}{Numeric vector consisting of the data}
}
\description{
Box-Cox transformation is only allowed for positive values. In
the case that the minimal value min one standard deviation is negative we add
this difference.
}

