% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateCovariates.R
\name{GetNormCovariate}
\alias{GetNormCovariate}
\title{Obtain new normalized covariates}
\usage{
GetNormCovariate(covariate, normCov)
}
\arguments{
\item{covariate}{The new value of the covariate}

\item{normCov}{The normalized covariate object}
}
\description{
Returns the normalized value of a new covariate value with respect to an
existing normCov object
}

