% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries.R
\name{IsLonLatPointInArea}
\alias{IsLonLatPointInArea}
\title{Wrapper for IsPointInArea for lat lon data}
\usage{
IsLonLatPointInArea(lon, lat, area)
}
\arguments{
\item{lon}{numeric vector of longitudes}

\item{lat}{numeric vector of latitudes}

\item{area}{Extending SpatialPolygons object}
}
\description{
Wrapper for IsPointInArea for lat lon data
}

