% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries.R
\name{IsPointInArea}
\alias{IsPointInArea}
\title{Determines wether spatial points are in area or not}
\usage{
IsPointInArea(points, area)
}
\arguments{
\item{points}{SpatialPoints object}

\item{area}{Extending SpatialPolygons object}
}
\value{
Logical vector with TRUE for points in area
}
\description{
Determines wether spatial points are in area or not
}

