% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateCovariates.R
\name{SetNormCovariate}
\alias{SetNormCovariate}
\title{Normalizes covariates}
\usage{
SetNormCovariate(covariate)
}
\arguments{
\item{covariate}{Covariate to be normalized}
}
\value{
list with three entries, \code{val} the normalized covariate,
 \code{meanCov} the mean of the covariate, and \code{sdCov} its standard
 deviation
}
\description{
Normalizes covariates
}

