% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateCovariates.R
\name{SetNormCovariate}
\alias{SetNormCovariate}
\title{Normalizes covariates}
\usage{
SetNormCovariate(covariate)
}
\arguments{
\item{covariate}{Covariate to be normalized}
}
\description{
Normalizes covariates
}

